code smell fixes

+++ test/Filtering/Test_FirFilter.cpp
@@ -16,7 +16,7 @@ using namespace dspbb;
 //------------------------------------------------------------------------------
 
 TimeSignal<float> GenTestSignal(size_t sampleRate, float frequency, float length = 1.0f) {
-	const size_t numSamples = std::max(size_t(1), size_t(sampleRate * double(length)));
+	const size_t numSamples = std::max(size_t(1), size_t(double(sampleRate) * double(length)));
 	return SineWave<float, TIME_DOMAIN>(numSamples, sampleRate, frequency);
 }
